Extract useful helper from AwsS3UrlReader to integration package
* Other integration expose similar functions so moving S3 to here also
