+++ .changeset/slow-sheep-admire.md
@@ -1,0 +1,6 @@
+---
+'@backstage/integration': minor
+'@backstage/backend-common': patch
+---
+
+Expose useful AWS S3 URL parsing functionality for URL integration.
